djangoproject/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
clipboard==0.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.0.1
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
natsort==8.2.0
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.1
Pillow==9.4.0
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.2
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
robotframework==6.0.1
robotframework-datadriver==1.6.1
robotframework-pabot==2.8.0
robotframework-pythonlibcore==4.0.0
robotframework-seleniumlibrary==6.0.0
robotframework-stacktrace==0.4.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wsproto==1.2.0
xlrd==2.0.1
